Updated spinner for simplicity
